feat: Add persistence subproject using Spring Data JPA (#3)
Add persistence `subproject` using Spring Data JPA:
- Add `CustomerEntity` which represents the database entity and uses jakarta annotations to define the table name and columns.
- Add `CustomerRepo` which extends `JpaRepository`. This allows us to easily define new methods to interact with the database.
- Add `CustomerRepoFace` which allows you to easily interact with the database without the need to know about `CustomerEntity`. This class will be responsible for converting the model class to an entity class before saving it to the database and vice versa when returning a customer model object.
- Add `CustomerInput` which will only contain exactly the fields necessary for creating a new customer. We can add more fields for customer creation here if necessary. We don't know the id of the customer before creation as it is generated when added to the database. This means we don't need to send the id when creating a customer.
- Add `CustomerDto` and `CustomerInputDto` which are the data transfer objects and they use the `@JsonCreator` annotation to ensure that FasterXML Jackson can serialize/deserialize the JSON.
- Update `ICustomerController` and `CustomerController`
  - Add method for creating a customer with `CustomerInputDto`
  - Add method for retrieving a customer given an UUID.
- Add `CustomerService` which has the purpose of containing business logic. The controller can interact with the service and the service can interact with the data repository.
- Enable Liquibase by removing `liquibase.enabled: false` in `application.yml`.
- Add Liquibase database changelog for the initial schema.

---

Now this means that there’s suddenly 5 classes which in some way represent a customer. This could be considered overkill but for a large project, this could easily be the norm. The purpose of this repository is simply to showcase how a project could be structured. 
- `CustomerEntity` - represents a customer in the database. Uses Jakarta to define table and column names. 
- `Customer` - Model class with pure java code and uses no dependencies.
- `CustomerInput` - Model class with pure java code and uses no dependencies.
- `CustomerDto` - Data transfer object that Jackson can serialize.
- `CustomerInputDto` - Data transfer object that Jackson can serialize.

The split between classes such as `Customer` and `CustomerInput` is because you don’t need the id of a customer to create one. The id is first generated once it has been added to the database. The benefit here is that we avoid having to send null values for an irrelevant id field when trying to create a new customer. This might be a lot of boilerplate code, but you should always consider what fits your specific needs. The very same functionality could easily be implemented with less customer classes. For this project I have simply decided to split it to separate concerns.