build: Use JRE instead of JDK for final Docker image
Use Eclipse Temurin JRE 21 instead of JDK 21 in Dockerfile. This means that the final Docker image for the application will be take up less space. The JDK is for development and we do not need to use it for built Docker images. The JRE is all that we need for running the application.